+++ docs/changelog.rst
@@ -11,6 +11,7 @@ ChangeLog
 
     - Add support for :attr:`factory.fuzzy.FuzzyInteger.step`, thanks to `ilya-pirogov <https://github.com/ilya-pirogov>`_ (:issue:`120`)
     - Add :meth:`~factory.django.mute_signals` decorator to temporarily disable some signals, thanks to `ilya-pirogov <https://github.com>`_ (:issue:`122`)
+    - Add :class:`~factory.fuzzy.FuzzyFloat` (:issue:`124`)
 
 .. _v2.3.1: